requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-htmx==1.14.0
django-taggit==3.1.0
factory-boy==3.2.1
Faker==18.3.4
flake8==6.0.0
Markdown==3.4.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
